Fix bad envvar syntax on GitHub Action
Fix bad environment variable syntax on GitHub Action. The result of this
incorrect syntax was the SLACK_TOKEN not being passed through correctly
which led to HTTP status 400 on curl calls to the Slack Webhook API.
This update fixes it.

Also add a debug line to the slack-notify.sh script so that the on
failure we can more easily find the command that caused the error.