Add user info env vars for annotated tags
Set the GIT_COMMITTER_NAME and EMAIL env vars to something so
that annotated tags can be created. These tags contain info about the
user creating the tag (much like a commit does) and git will refuse to
create the tag if this information is not available.

We use a bogus email address - I mean, who uses email anymore?